Revert "Float-default (#72)"
This reverts commit fd969cd87f5369474ef7e013df3335746c100cd1.

+++ tests/config/test_config.py
@@ -18,7 +18,7 @@ def test_set_and_get_default_spacing():
     spacing = (10e-6, 10e-6)
     set_default_spacing(spacing)
     result = get_default_spacing()
-    assert torch.allclose(result, torch.tensor(spacing))
+    assert torch.allclose(result, torch.tensor(spacing, dtype=torch.double))
 
 
 def test_get_default_wavelength_not_set():
@@ -30,4 +30,4 @@ def test_set_and_get_default_wavelength():
     wavelength = 700e-6
     set_default_wavelength(wavelength)
     result = get_default_wavelength()
-    assert torch.allclose(result, torch.tensor(wavelength))
+    assert torch.allclose(result, torch.tensor(wavelength, dtype=torch.double))
